Fix breaking tests
